+++ Tests/StreamChatTests/StreamChat/ChatClient_Tests.swift
@@ -319,7 +319,7 @@ final class ChatClient_Tests: XCTestCase {
         // Set a connection Id waiter and assert it's `nil`
         var providedConnectionId: ConnectionId?
         client.provideConnectionId {
-            providedConnectionId = $0
+            providedConnectionId = $0.value
         }
         XCTAssertNil(providedConnectionId)
         
@@ -337,7 +337,7 @@ final class ChatClient_Tests: XCTestCase {
             .webSocketClient(testEnv.webSocketClient!, didUpdateConnectionState: .connected(connectionId: connectionId))
         
         AssertAsync.willBeEqual(providedConnectionId, connectionId)
-        XCTAssertEqual(try waitFor { client.provideConnectionId(completion: $0) }, connectionId)
+        XCTAssertEqual(try waitFor { client.provideConnectionId(completion: $0) }.value, connectionId)
         
         // Simulate WebSocketConnection disconnecting and assert connectionId is reset
         testEnv.webSocketClient?.connectionStateDelegate?
@@ -345,7 +345,7 @@ final class ChatClient_Tests: XCTestCase {
         
         providedConnectionId = nil
         client.provideConnectionId {
-            providedConnectionId = $0
+            providedConnectionId = $0.value
         }
         AssertAsync.staysTrue(providedConnectionId == nil)
     }
@@ -365,7 +365,7 @@ final class ChatClient_Tests: XCTestCase {
         // Set a connection Id waiter and set `providedConnectionId` to a non-nil value
         var providedConnectionId: ConnectionId? = .unique
         client.provideConnectionId {
-            providedConnectionId = $0
+            providedConnectionId = $0.value
         }
         XCTAssertNotNil(providedConnectionId)
         
@@ -462,7 +462,7 @@ final class ChatClient_Tests: XCTestCase {
             // This is to simulate the case where 2 entities call this func
             // Like, calling `synchronize` in a delegate callback
             client.provideConnectionId {
-                providedConnectionId = $0
+                providedConnectionId = $0.value
             }
         }
         XCTAssertNil(providedConnectionId)
@@ -834,6 +834,183 @@ final class ChatClient_Tests: XCTestCase {
         }
     }
 
+    // MARK: Provide token
+
+    func test_provideToken_returnsValue_whenAlreadyHasToken() {
+        let client = ChatClient(config: inMemoryStorageConfig, environment: testEnv.environment)
+
+        let existingToken = Token.unique()
+        client.setToken(token: existingToken)
+
+        var result: Result<Token, Error>?
+        let expectation = self.expectation(description: "Provide Connection Id Completion")
+        client.provideToken(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+
+        // Sync execution
+        XCTAssertNotNil(result)
+
+        // Force timeout
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertEqual(result?.value, existingToken)
+    }
+
+    func test_provideToken_returnsErrorOnTimeout() {
+        let client = ChatClient(config: inMemoryStorageConfig, environment: testEnv.environment)
+
+        var result: Result<Token, Error>?
+        let expectation = self.expectation(description: "Provide Token Completion")
+        client.provideToken(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+        XCTAssertNil(result)
+
+        // Force timeout
+        VirtualTimeTimer.time.run()
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertTrue(result?.error is ClientError.WaiterTimeout)
+    }
+
+    func test_provideToken_returnsErrorOnMissingValue() {
+        let client = ChatClient(config: inMemoryStorageConfig, environment: testEnv.environment)
+
+        var result: Result<Token, Error>?
+        let expectation = self.expectation(description: "Provide Token Completion")
+        client.provideToken(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+        XCTAssertNil(result)
+
+        // Complete with nil
+        client.completeTokenWaiters(token: nil)
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertTrue(result?.error is ClientError.MissingToken)
+    }
+
+    func test_provideToken_returnsValue_whenCompletingTokenWaiters() {
+        let client = ChatClient(config: inMemoryStorageConfig, environment: testEnv.environment)
+
+        var result: Result<Token, Error>?
+        let expectation = self.expectation(description: "Provide Token Completion")
+        client.provideToken(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+        XCTAssertNil(result)
+
+        // Complete with token
+        let token = Token.unique()
+        client.completeTokenWaiters(token: token)
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertEqual(result?.value, token)
+    }
+
+    // MARK: Provide connection id
+
+    func test_provideConnectionId_returnsErrorOnActiveModeClient() {
+        let client = ChatClient(config: inactiveInMemoryStorageConfig, environment: testEnv.environment)
+
+        var result: Result<ConnectionId, Error>?
+        let expectation = self.expectation(description: "Provide Connection Id Completion")
+        client.provideConnectionId(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+
+        // Sync execution
+        XCTAssertNotNil(result)
+
+        // Force timeout
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertTrue(result?.error is ClientError.ClientIsNotInActiveMode)
+    }
+
+    func test_provideConnectionId_returnsValue_whenAlreadyHasConnectionId() {
+        let client = ChatClient(config: inMemoryStorageConfig, environment: testEnv.environment)
+
+        let existingConnectionId = "connection-id"
+        client.webSocketClient(client.mockWebSocketClient, didUpdateConnectionState: .connected(connectionId: existingConnectionId))
+
+        var result: Result<ConnectionId, Error>?
+        let expectation = self.expectation(description: "Provide Connection Id Completion")
+        client.provideConnectionId(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+
+        // Sync execution
+        XCTAssertNotNil(result)
+
+        // Force timeout
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertEqual(result?.value, existingConnectionId)
+    }
+
+    func test_provideConnectionId_returnsErrorOnTimeout() {
+        let client = ChatClient(config: inMemoryStorageConfig, environment: testEnv.environment)
+
+        var result: Result<ConnectionId, Error>?
+        let expectation = self.expectation(description: "Provide Connection Id Completion")
+        client.provideConnectionId(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+        XCTAssertNil(result)
+
+        // Force timeout
+        VirtualTimeTimer.time.run()
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertTrue(result?.error is ClientError.WaiterTimeout)
+    }
+
+    func test_provideConnectionId_returnsErrorOnMissingValue() {
+        let client = ChatClient(config: inMemoryStorageConfig, environment: testEnv.environment)
+
+        var result: Result<ConnectionId, Error>?
+        let expectation = self.expectation(description: "Provide Connection Id Completion")
+        client.provideConnectionId(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+        XCTAssertNil(result)
+
+        // Complete with nil
+        client.completeConnectionIdWaiters(connectionId: nil)
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertTrue(result?.error is ClientError.MissingConnectionId)
+    }
+
+    func test_provideConnectionId_returnsValue_whenCompletingConnectionIdWaiters() {
+        let client = ChatClient(config: inMemoryStorageConfig, environment: testEnv.environment)
+
+        var result: Result<ConnectionId, Error>?
+        let expectation = self.expectation(description: "Provide Connection Id Completion")
+        client.provideConnectionId(timeout: 0.1) {
+            result = $0
+            expectation.fulfill()
+        }
+        XCTAssertNil(result)
+
+        // Complete with connection id
+        let connectionId = "connection-id"
+        client.completeConnectionIdWaiters(connectionId: connectionId)
+        waitForExpectations(timeout: 0.1)
+
+        XCTAssertEqual(result?.value, connectionId)
+    }
+
     // MARK: Set token
 
     func test_setToken_savesToken_completesWaiters() throws {
@@ -846,8 +1023,8 @@ final class ChatClient_Tests: XCTestCase {
 
         let expectation = self.expectation(description: "Token waiter")
         var providedToken: Token?
-        client.provideToken { token in
-            providedToken = token
+        client.provideToken { result in
+            providedToken = result.value
             expectation.fulfill()
         }
 
@@ -875,8 +1052,8 @@ final class ChatClient_Tests: XCTestCase {
 
         let expectation = self.expectation(description: "Token waiter")
         var providedToken: Token?
-        client.provideToken { token in
-            providedToken = token
+        client.provideToken { result in
+            providedToken = result.value
             expectation.fulfill()
         }
 
@@ -901,8 +1078,8 @@ final class ChatClient_Tests: XCTestCase {
         XCTAssertNil(client.currentUserId)
         XCTAssertEqual(client.backgroundWorkers.count, 0)
 
-        client.provideToken { token in
-            if token != nil {
+        client.provideToken { result in
+            if result.value != nil {
                 XCTFail("Should not complete waiters")
             }
         }
@@ -997,8 +1174,8 @@ final class ChatClient_Tests: XCTestCase {
 
         let expectation = self.expectation(description: "Complete existing waiter")
         var providedToken: Token?
-        client.provideToken { token in
-            providedToken = token
+        client.provideToken { result in
+            providedToken = result.value
             expectation.fulfill()
         }
 
@@ -1011,8 +1188,8 @@ final class ChatClient_Tests: XCTestCase {
 
         client.logout()
 
-        client.provideToken { token in
-            if token != nil {
+        client.provideToken { result in
+            if result.value != nil {
                 XCTFail("Should not complete waiters for new token")
             }
         }
@@ -1050,7 +1227,7 @@ final class ChatClient_Tests: XCTestCase {
         var providedConnectionId: ConnectionId? = .unique
         var connectionIdCallbackCalled = false
         client.provideConnectionId {
-            providedConnectionId = $0
+            providedConnectionId = $0.value
             connectionIdCallbackCalled = true
         }